Working on getupdatedtaxfor a user

+++ source/Devon4Net.WebAPI.Implementation/Business/InhabitantsManagement/Dto/AsignTaxToUserDto.cs
@@ -6,7 +6,7 @@
         public string Surname { get; set; }
         public string TaxName { get; set; }
         public int TaxYear { get; set; }
-        public int BaseAmount { get; set; }
+        public double BaseAmount { get; set; }
         public string Reference { get; set; }
     }
 }
